Move db support to its own package
