+++ app/src/main/groovy/lt/vilnius/tvarkau/support/db/LiquibaseSupport.groovy
@@ -1,10 +1,11 @@
-package lt.vilnius.tvarkau.support
+package lt.vilnius.tvarkau.support.db
 
 import liquibase.Contexts
 import liquibase.Liquibase
 import liquibase.database.DatabaseFactory
 import liquibase.database.jvm.JdbcConnection
 import liquibase.resource.ClassLoaderResourceAccessor
+import lt.vilnius.tvarkau.support.db.DatabaseSupport
 
 import java.sql.Connection
 
@@ -14,7 +15,7 @@ class LiquibaseSupport {
         Connection conn
         Liquibase liquibase
         try {
-            conn = DataSourceSupport.getConnection()
+            conn = DatabaseSupport.getConnection()
             def database = DatabaseFactory.instance.findCorrectDatabaseImplementation(new JdbcConnection(conn))
             liquibase = new Liquibase('db/changelog.xml', new ClassLoaderResourceAccessor(), database)
             liquibase.update(new Contexts())
